pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sft-grpo"
version = "0.1.0"
description = "SFT experiments for Mistral 7B and Llama 3B (model-centric structure)"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]

dependencies = [
    "torch>=2.0.0",
    "transformers>=4.35.0",
    "datasets>=2.14.0",
    "peft>=0.6.0",
    "python-dotenv>=1.0.0",
    "accelerate>=0.24.0",
    "pandas>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "ipykernel>=6.0.0",
    "jupyter>=1.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["sft_grpo*"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']

[tool.ruff]
line-length = 100
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["src/sft_grpo/sft/mistral/tests", "src/sft_grpo/sft/llama/tests"]
pythonpath = ["src"]